+++ commandline/src/main/java/org/prgrms/kdt/util/ControllerResult.java
@@ -5,7 +5,7 @@ public class ControllerResult {
     private ControllerResult() {
     }
 
-    public static String OK = "ok";
+    public static String SUCCESS_RESULT = "ok";
 
     public static int USER_ERROR = 404;
     public static int SERVER_ERROR = 500;
